Move useDetectWindowSize
